Refactor queue test
